debrid-downloader/pyproject.toml
[tool.poetry]
name = "debrid-downloader"
version = "0.1.0"
description = ""
authors = ["Ken Schiano <kschiano3@gatech.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
pyyaml = "^6.0.1"
peewee = "^3.17.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
